--- iot-e2e-tests/android/app/src/androidTest/java/com/microsoft/azure/sdk/iot/android/iothub/errorinjection/methods/DirectMethodsErrInjAndroidRunner.java
@@ -1,27 +1,0 @@
-/*
- *  Copyright (c) Microsoft. All rights reserved.
- *  Licensed under the MIT license. See LICENSE file in the project root for full license information.
- */
-
-package com.microsoft.azure.sdk.iot.android.iothub.errorinjection.methods;
-
-import com.microsoft.azure.sdk.iot.android.helper.TestGroup5;
-import com.microsoft.azure.sdk.iot.device.IotHubClientProtocol;
-import com.microsoft.azure.sdk.iot.service.auth.AuthenticationType;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-
-import tests.integration.com.microsoft.azure.sdk.iot.helpers.ClientType;
-import tests.integration.com.microsoft.azure.sdk.iot.iothub.errorinjection.DirectMethodsErrInjTests;
-
-@TestGroup5
-@RunWith(Parameterized.class)
-public class DirectMethodsErrInjAndroidRunner extends DirectMethodsErrInjTests
-{
-    public DirectMethodsErrInjAndroidRunner(IotHubClientProtocol protocol, AuthenticationType authenticationType, ClientType clientType) throws Exception
-    {
-        super(protocol, authenticationType, clientType);
-    }
-}
-
