+++ iot-e2e-tests/android/app/src/androidTest/java/com/microsoft/azure/sdk/iot/android/iothub/serviceclient/ConfigurationsClientAndroidRunner.java
@@ -1,0 +1,14 @@
+/*
+ *  Copyright (c) Microsoft. All rights reserved.
+ *  Licensed under the MIT license. See LICENSE file in the project root for full license information.
+ */
+
+package com.microsoft.azure.sdk.iot.android.iothub.serviceclient;
+
+import com.microsoft.azure.sdk.iot.android.helper.TestGroup6;
+import tests.integration.com.microsoft.azure.sdk.iot.iothub.serviceclient.ConfigurationsClientTests;
+
+@TestGroup6
+public class ConfigurationsClientAndroidRunner extends ConfigurationsClientTests
+{
+}
