+++ iot-e2e-tests/android/app/src/androidTest/java/com/microsoft/azure/sdk/iot/android/iothub/serviceclient/TwinClientAndroidRunner.java
@@ -6,10 +6,9 @@
 package com.microsoft.azure.sdk.iot.android.iothub.serviceclient;
 
 import com.microsoft.azure.sdk.iot.android.helper.TestGroup7;
-
-import tests.integration.com.microsoft.azure.sdk.iot.iothub.serviceclient.RegistryClientTests;
+import tests.integration.com.microsoft.azure.sdk.iot.iothub.serviceclient.TwinClientTests;
 
 @TestGroup7
-public class RegistryClientAndroidRunner extends RegistryClientTests
+public class TwinClientAndroidRunner extends TwinClientTests
 {
 }
