+++ iot-e2e-tests/android/app/src/androidTest/java/com/microsoft/azure/sdk/iot/android/iothub/serviceclient/TwinClientPnPAndroidRunner.java
@@ -3,25 +3,21 @@
  *  Licensed under the MIT license. See LICENSE file in the project root for full license information.
  */
 
-package com.microsoft.azure.sdk.iot.android.iothub.errorinjection.twin;
+package com.microsoft.azure.sdk.iot.android.iothub.serviceclient;
 
-import com.microsoft.azure.sdk.iot.android.helper.TestGroup7;
+import com.microsoft.azure.sdk.iot.android.helper.TestGroup14;
 import com.microsoft.azure.sdk.iot.device.IotHubClientProtocol;
 import com.microsoft.azure.sdk.iot.service.auth.AuthenticationType;
-
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-
-import java.io.IOException;
-
 import tests.integration.com.microsoft.azure.sdk.iot.helpers.ClientType;
-import tests.integration.com.microsoft.azure.sdk.iot.iothub.errorinjection.GetTwinErrInjTests;
+import tests.integration.com.microsoft.azure.sdk.iot.iothub.serviceclient.TwinClientPnPTests;
 
-@TestGroup7
+@TestGroup14
 @RunWith(Parameterized.class)
-public class GetTwinErrInjAndroidRunner extends GetTwinErrInjTests
+public class TwinClientPnPAndroidRunner extends TwinClientPnPTests
 {
-    public GetTwinErrInjAndroidRunner(IotHubClientProtocol protocol, AuthenticationType authenticationType, ClientType clientType) throws IOException
+    public TwinClientPnPAndroidRunner(IotHubClientProtocol protocol, AuthenticationType authenticationType, ClientType clientType)
     {
         super(protocol, authenticationType, clientType);
     }
