--- iot-e2e-tests/android/app/src/androidTest/java/com/microsoft/azure/sdk/iot/android/iothub/twin/ReportedPropertiesAndroidRunner.java
@@ -1,28 +1,0 @@
-/*
- *  Copyright (c) Microsoft. All rights reserved.
- *  Licensed under the MIT license. See LICENSE file in the project root for full license information.
- */
-
-package com.microsoft.azure.sdk.iot.android.iothub.twin;
-
-import com.microsoft.azure.sdk.iot.android.helper.TestGroup15;
-import com.microsoft.azure.sdk.iot.device.IotHubClientProtocol;
-import com.microsoft.azure.sdk.iot.service.auth.AuthenticationType;
-
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
-
-import java.io.IOException;
-
-import tests.integration.com.microsoft.azure.sdk.iot.helpers.ClientType;
-import tests.integration.com.microsoft.azure.sdk.iot.iothub.twin.ReportedPropertiesTests;
-
-@TestGroup15
-@RunWith(Parameterized.class)
-public class ReportedPropertiesAndroidRunner extends ReportedPropertiesTests
-{
-    public ReportedPropertiesAndroidRunner(IotHubClientProtocol protocol, AuthenticationType authenticationType, ClientType clientType) throws IOException
-    {
-        super(protocol, authenticationType, clientType);
-    }
-}
