+++ iot-e2e-tests/common/src/test/java/tests/integration/com/microsoft/azure/sdk/iot/iothub/serviceclient/ConfigurationsClientTests.java
@@ -65,8 +65,8 @@ public class ConfigurationsClientTests extends IntegrationTest
         public String deviceId;
         public String moduleId;
         public String configId;
-        private ConfigurationsClient configurationsClient;
-        private RegistryClient registryClient;
+        private final ConfigurationsClient configurationsClient;
+        private final RegistryClient registryClient;
 
         public ConfigurationsClientTestInstance()
         {
